[MRI Violations] Improved UI
